# docs/openapi/mobile_api.yaml
openapi: 3.1.0
info:
  title: Ratatoskr Mobile API
  description: RESTful API for Android/iOS mobile clients
  version: 1.0.0
servers:
  - url: https://ratatoskrapi.po4yka.com
    description: Production
  - url: https://staging-ratatoskrapi.po4yka.com
    description: Staging
  - url: http://localhost:8000
    description: Local development
paths:
...
  /v1/webhooks/:
    get:
      tags:
      - Webhooks
      summary: List Webhook Subscriptions
      description: List user's webhook subscriptions.
      operationId: list_subscriptions_v1_webhooks_get
      security:
      - HTTPBearer: []
      responses:
        '200':
          description: List of webhook subscriptions
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
      - Webhooks
      summary: Create Webhook Subscription
      description: Create a new webhook subscription.
      operationId: create_subscription_v1_webhooks_post
      security:
      - HTTPBearer: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                url:
                  type: string
                events:
                  type: array
                  items:
                    type: string
              required:
              - url
              - events
        required: true
      responses:
        '201':
          description: Webhook subscription created
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/webhooks/{webhook_id}:
    get:
      tags:
      - Webhooks
      summary: Get Webhook Subscription
      description: Get a webhook subscription's details.
      operationId: get_subscription_v1_webhooks__webhook_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Webhook subscription details
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
      - Webhooks
      summary: Update Webhook Subscription
      description: Update a webhook subscription.
      operationId: update_subscription_v1_webhooks__webhook_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                url:
                  type: string
                events:
                  type: array
                  items:
                    type: string
                enabled:
                  type: boolean
        required: true
      responses:
        '200':
          description: Webhook subscription updated
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
      - Webhooks
      summary: Delete Webhook Subscription
      description: Soft-delete a webhook subscription.
      operationId: delete_subscription_v1_webhooks__webhook_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Webhook subscription deleted
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/webhooks/{webhook_id}/rotate-secret:
    post:
      tags:
      - Webhooks
      summary: Rotate Webhook Secret
      description: Generate a new secret for the subscription. Returns the new secret once.
      operationId: rotate_secret_v1_webhooks__webhook_id__rotate_secret_post
      security:
      - HTTPBearer: []
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: New secret generated
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/webhooks/{webhook_id}/test:
    post:
      tags:
      - Webhooks
      summary: Send Test Webhook
      description: Send a test event to the webhook URL and return the delivery result.
      operationId: send_test_webhook_v1_webhooks__webhook_id__test_post
      security:
      - HTTPBearer: []
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Test delivery result
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  # -----------------------------------------------------------------------
  # Rules
  # -----------------------------------------------------------------------
...
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
  responses:
...
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Meta:
      type: object
      description: Response metadata for observability and pagination.
      required:
      - correlation_id
      - timestamp
      - version
      - api_version
      properties:
        correlation_id:
          type: string
          description: Request correlation ID for debugging (always present, may be empty string).
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 response timestamp.
        version:
          type: string
          description: App / build version (changes every deploy). Sourced from APP_VERSION env var.
        api_version:
          type: string
          description: |
            API contract semver. Distinct from `version` — bumps only on breaking contract
            changes (response shape, request shape, routing surface). Mobile / CLI clients
            should pin against this value and refuse to upgrade major versions blindly.
        build:
          type: string
          nullable: true
          description: Build identifier (if available).
        pagination:
          $ref: '#/components/schemas/Pagination'
        debug:
          type: object
          nullable: true
          description: Debug information (only in development mode).
    Pagination:
      type: object
      description: Pagination metadata for list responses.
      required:
      - total
      - limit
      - offset
      - hasMore
      properties:
        total:
          type: integer
          description: Total number of items matching the query.
        limit:
          type: integer
          description: Maximum items per page.
        offset:
          type: integer
          description: Current offset from start.
        hasMore:
          type: boolean
          description: Whether more items are available.
        nextCursor:
          type: integer
          nullable: true
          description: Cursor for next page (if using cursor-based pagination).
...
    ErrorObject:
      type: object
      description: Structured error details for programmatic handling.
      required:
      - code
      - errorType
      - message
      - retryable
      - correlation_id
      properties:
        code:
          type: string
          description: Machine-readable error code for programmatic handling.
          enum:
          # Client errors
          - VALIDATION_ERROR
          - UNAUTHORIZED
          - FORBIDDEN
          - NOT_FOUND
          - CONFLICT
          - RATE_LIMIT_EXCEEDED
          - SESSION_EXPIRED
          # Token/Auth errors
          - TOKEN_EXPIRED
          - TOKEN_INVALID
          - TOKEN_REVOKED
          - TOKEN_WRONG_TYPE
          - REFRESH_RATE_LIMITED
          # Sync errors
          - SYNC_SESSION_EXPIRED
          - SYNC_SESSION_NOT_FOUND
          - SYNC_SESSION_FORBIDDEN
          - SYNC_NO_CHANGES
          - SYNC_CONFLICT
          - SYNC_INVALID_ENTITY
          - SYNC_ENTITY_NOT_FOUND
          # Server errors
          - INTERNAL_ERROR
          - DATABASE_ERROR
          - EXTERNAL_API_ERROR
          - PROCESSING_ERROR
          - AUTH_SERVICE_UNAVAILABLE
          # Configuration errors
          - CONFIGURATION_ERROR
          - FEATURE_DISABLED
        errorType:
          type: string
          description: Error category for client-side handling logic.
          enum:
          - authentication
          - authorization
          - validation
          - not_found
          - conflict
          - rate_limit
          - external_service
          - sync
          - internal
          - configuration
        message:
          type: string
          description: Human-readable error message.
        retryable:
          type: boolean
          description: Whether the client should retry the request.
          default: false
        details:
          type: object
          nullable: true
          description: Additional error context (field errors, resource IDs, etc.).
        correlation_id:
          type: string
          description: Request correlation ID for debugging (always present).
        retry_after:
          type: integer
          nullable: true
          description: Seconds to wait before retrying (for rate limit errors).
    ErrorResponse:
      type: object
      required:
      - success
      - error
      - meta
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
...